# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API TESIS - Sistema de Clasificación de Riesgo",
        "description": "Documentación de la API del sistema de tesis",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/api/permissions": {
            "get": {
                "tags": [
                    "Permisos"
                ],
                "summary": "Listar permisos",
                "description": "Lista permisos activos, inactivos y/o eliminados lógicamente según filtros",
                "operationId": "27bb1606a7b293d1d49f57bb94ae71f7",
                "parameters": [
                    {
                        "name": "estado",
                        "in": "query",
                        "description": "Filtrar por estado (1=activo, 0=inactivo)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "withTrashed",
                        "in": "query",
                        "description": "Incluir registros eliminados lógicamente",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de permisos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Permisos"
                ],
                "summary": "Crear permiso",
                "operationId": "37c918c86b5085b0eb088fa760c7175d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "editar-usuarios"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Permiso creado correctamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/permissions/{id}": {
            "get": {
                "tags": [
                    "Permisos"
                ],
                "summary": "Mostrar permiso",
                "description": "Devuelve un permiso específico (incluye soft deleted)",
                "operationId": "397983813683c0c602fc1a036992367a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permiso encontrado"
                    },
                    "404": {
                        "description": "Permiso no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Permisos"
                ],
                "summary": "Actualizar permiso",
                "operationId": "db11d4d00da3ece5b99187df6f1e86e7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "crear-reportes"
                                    },
                                    "estado": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ],
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permiso actualizado correctamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Permisos"
                ],
                "summary": "Eliminar permiso",
                "description": "Elimina un permiso mediante soft delete",
                "operationId": "f0afbd22579524a5bb83b653f7dd18dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permiso eliminado (soft delete)"
                    }
                }
            }
        },
        "/api/permissions/{id}/deactivate": {
            "patch": {
                "tags": [
                    "Permisos"
                ],
                "summary": "Desactivar permiso",
                "operationId": "f6be9f4f8c0dee0d88b6ee8663cf64dd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permiso desactivado"
                    }
                }
            }
        },
        "/api/permissions/{id}/activate": {
            "patch": {
                "tags": [
                    "Permisos"
                ],
                "summary": "Activar permiso",
                "operationId": "371bf50d4c7e8a33883b358f53f42895",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permiso activado"
                    }
                }
            }
        },
        "/api/permissions/{id}/restore": {
            "patch": {
                "tags": [
                    "Permisos"
                ],
                "summary": "Restaurar permiso",
                "operationId": "a4d15ee25aa5243dee9ba290bfa77753",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permiso restaurado correctamente"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Ingresa tu token de autenticación Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}